chaindb: move chaindb records to records file.
