+++ lib/blockchain/chain.js
@@ -943,7 +943,7 @@ class Chain extends AsyncEmitter {
    * Verify a renewal.
    * @param {Hash} hash
    * @param {Number} height
-   * @returns {Boolean}
+   * @returns {Promise<Boolean>}
    */
 
   async verifyRenewal(hash, height) {
